[level 3] Title: 아방가르드 타일링, Time: 8.64 ms, Memory: 78.5 MB -BaekjoonHub

+++ 프로그래머스/3/181186. 아방가르드 타일링/아방가르드 타일링.java
@@ -25,6 +25,6 @@ class Solution {
             dp[i] = (dp[i-1] + dp[i-2]*2 + dp[i-3]*5 + dp2[i-4]*2 + dp2[i-5]*2 + dp2[i-6]*4)%MOD;
             dp2[i] = (dp[i] + dp2[i-3])%MOD;
         }
-        return Math.toIntExact(dp[n]);
+        return (int)dp[n];
     }
 }
